+++ packages/expo-brownfield/plugin/src/ios/plugins/withPodfilePropertiesPlugin.ts
@@ -4,8 +4,9 @@ import type { PluginConfig } from '../types';
 
 const withPodfilePropertiesPlugin: ConfigPlugin<PluginConfig> = (config, pluginConfig) => {
   return withPodfileProperties(config, (config) => {
-    config.modResults['ios.useFrameworks'] = 'static';
-    config.modResults['ios.brownfieldTargetName'] = pluginConfig.targetName;
+    if (!pluginConfig.usePrebuiltReactNative) {
+      config.modResults['ios.useFrameworks'] = 'static';
+    }
     return config;
   });
 };